Rfc2898DeriveBytes As per MSDN: 'Implements password-based key derivation functionality, PBKDF2, by using a pseudo-random number generator based on HMACSHA1.'
Now, I know 1000 iterations is the default but for in-memory encryption that simply is far too many; it would make the whole experience far too sluggish and as far as on-disk encryption is concerned; 1000 or more may happen soon but first I need to create a splashscreen of sorts to inform the user of the delay and prevent them from accessing the UI before loading has completed.
Additionally, I would need to change the way PinCushion saves its data; as it stands, it saves the data with every single form of manipulation and with every single object having its own key, IV and salt that does tend to slow things down a lot with Rfc2898DeriveBytes.
So, for the time being, 10 iterations in-memory and 100 on-disk.