Re-introduce own in-memory encryption plus... ... renaming of the lists (Profile.Services and Service.Accounts).
The reason for reintroducing my own in-memory encryption is because after some experimentation I found that NetworkCredential seemingly internally does NOT use only SecureStrings as I was led to believe previously. To rule it out either way, I decided to go back to my own in-memory encryption and consequently, to add a fast_encrypt/decrypt parameter to the Encryt/Decrypt methods. All in all, it does add a little to the loading time but not too much. About 50ms on my ~70 account personal data file.